requirements.txt
django
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
python-dotenv
pydantic
requests
twilio
